Hook up the agent connection store and separate thread archive view
